pyproject.toml
[project]
name = "pdf-outline-edit"
version = "24.2.3"
authors = [{ name = "slbtty", email = "shenlebantongying@gmail.com" }]
description = "Mini PDF outline editor"
readme = "README.md"
requires-python = ">=3.11"
classifiers = ["Programming Language :: Python :: 3"]

dependencies = ["pymupdf"]

[project.optional-dependencies]
qt = ["PySide6-Essentials>=6.6"]

[project.gui-scripts]
pdf-outline-edit = "pdfoe.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
